Refactor some components to pages
